Preserve interpreted options when input to compiler is already-built FileDescriptorProto (#217)
When a resolver provides an input to the compiler in the form of a
`*descriptorpb.FileDescriptorProto` (instead of as source code, or as a
fully-baked `protoreflect.FileDescriptor`), the step that interprets
options could inadvertently clobber field options (thinking it was the
case that the options were empty, just because there were no
uninterpreted options to process). Also, the
`linker.Result.CanonicalProto()` method would also clobber options, even
if it had no canonical serialized bytes to replace them. This commit
fixes both of these issues.