Use fewer allocations when building descriptors in the linker package (#290)
This uses many fewer, larger allocations, to allocate all of the descendant
descriptors in a file in as few chunks as possible (by allocating slices of
flattened structs).

This change also eliminates many allocations related to computing
fully-qualified-names by moving the construction of the descriptor hierarchy
(which includes computing and storing full names) to *before* we import
the names into the symbol table (which previously had to compute/allocate
the full names twice: once when checking for collisions, and again in a
second pass to store the names when no collisions were found).

(cherry picked from commit 93923d2711547360a76689c6a402c39a68e2ce4b)