Add final checks that were added to protoc between v27.0-rc3 and final v27.0 (#309)
This change mirrors two of the changes in the final v27.0 of protoc
that were not previously implemented in this compiler:

1. A new `feature_support` field on `EnumValueOptions` to allow defining
the lifetime of a feature value. This is similar to the field of the
same name and type on `FieldOptions`, but it controls the actual enum
values and in which editions they are valid.
2. A new check that a feature is not used from the same file in which
it's defined.

This change updates this repo to use the latest (final) v27.0 release and
also updates the protobuf-go dependency to include the v27.0 version
of the `descriptorpb` package.

(cherry picked from commit 4190af10a5e22d8764ccca70ad62d23f56b978c2)